* online-interface/xsparqlrewrite: show stderr msgs in interface
* software/xsparqlow/xsparql/debug.py: aux function to print debug
info

* software/xsparqlow/xsparql/grammar.py: Error printing is now done to
stderr.  auxiliary functions moved to a xquery module available at
http://xsparql.deri.org/xsparql.xquery.

* software/xsparqlow/xsparql/lifrewriter.py: rewriting for the lifting
checks if it the variable is a valid subject, predicate or object.

* software/xsparqlow/xsparql/grammar.py: rewriting for the lifting
checks if it the variable is a valid subject, predicate or object.

* software/xsparqlow/xsparql/grammar.py: new syntax for nested
expressions in constructs: _:{} for a bnode, <{}> or {}:{} for
iri. The {} is defined as being a literal "{}".